Rename "Cheats_GiveMaxLives" to "Cheats_GiveAllEmblems"
The address 0x1a5577c is incorrectly labeled. The cheat does not give you 99 lives. Instead, it sets all the relevant flags to get all the emblems. The address is read at the beginning of the emblem count update function, so the cheat doesn't have any effect until you do something to update the emblem counter (exit a level).